Add a global variable resolver for rhai
also:

 - do some basic re-organising of the rhai code
 - update documentation and examples to use new "constants"
