+++ apollo-router/src/plugins/rhai/execution.rs
@@ -1,0 +1,5 @@
+//! execution module
+
+pub(crate) use crate::services::execution::*;
+pub(crate) type Response = super::RhaiExecutionResponse;
+pub(crate) type DeferredResponse = super::RhaiExecutionDeferredResponse;
