+++ apollo-router/src/plugins/rhai/mod.rs
@@ -31,6 +31,7 @@ use notify::EventKind;
 use notify::PollWatcher;
 use notify::RecursiveMode;
 use notify::Watcher;
+use once_cell::sync::Lazy;
 use rhai::module_resolvers::FileModuleResolver;
 use rhai::plugin::*;
 use rhai::serde::from_dynamic;
@@ -82,6 +83,9 @@ type SharedMut<T> = rhai::Shared<Mutex<Option<T>>>;
 
 pub(crate) const RHAI_SPAN_NAME: &str = "rhai_plugin";
 
+static RHAI_GLOBAL_VARIABLES: Lazy<Arc<Mutex<Map>>> =
+    Lazy::new(|| Arc::new(Mutex::new(Map::new())));
+
 impl<T> OptionDance<T> for SharedMut<T> {
     fn with_mut<R>(&self, f: impl FnOnce(&mut T) -> R) -> R {
         let mut guard = self.lock().expect("poisoned mutex");
@@ -105,24 +109,14 @@ impl<T> OptionDance<T> for SharedMut<T> {
     }
 }
 
-mod supergraph {
-    pub(crate) use crate::services::supergraph::*;
-    pub(crate) type Response = super::RhaiSupergraphResponse;
-    pub(crate) type DeferredResponse = super::RhaiSupergraphDeferredResponse;
-}
+mod execution;
+mod subgraph;
+mod supergraph;
 
-mod execution {
-    pub(crate) use crate::services::execution::*;
-    pub(crate) type Response = super::RhaiExecutionResponse;
-    pub(crate) type DeferredResponse = super::RhaiExecutionDeferredResponse;
-}
-
-mod subgraph {
-    pub(crate) use crate::services::subgraph::*;
-}
-
+// We have to keep the modules that we export using `export_module` inline because
+// error[E0658]: non-inline modules in proc macro input are unstable
 #[export_module]
-mod router_base64_mod {
+mod router_base64 {
     #[rhai_fn(pure, return_raw)]
     pub(crate) fn decode(input: &mut ImmutableString) -> Result<String, Box<EvalAltResult>> {
         String::from_utf8(base64::decode(input.as_bytes()).map_err(|e| e.to_string())?)
@@ -135,8 +129,10 @@ mod router_base64_mod {
     }
 }
 
+// We have to keep the modules that we export using `export_module` inline because
+// error[E0658]: non-inline modules in proc macro input are unstable
 #[export_module]
-mod router_plugin_mod {
+mod router_plugin {
     // It would be nice to generate get_originating_headers and
     // set_originating_headers for all response types.
     // However, variations in the composition
@@ -400,12 +396,20 @@ impl EngineBlock {
         let mut scope = Scope::new();
         // Keep these two lower cases ones as mistakes until 2.0
         // At 2.0 (or maybe before), replace with upper case
+        // Note: Any constants that we add to scope here, *must* be catered for in the on_var
+        // functionality in `new_rhai_engine`.
         scope.push_constant("apollo_sdl", sdl.to_string());
         scope.push_constant("apollo_start", Instant::now());
 
-        scope.push_constant(
-            "APOLLO_AUTHENTICATION_JWT_CLAIMS",
-            "apollo_authentication::JWT::claims".to_string(),
+        // Since constants in Rhai don't give us the behaviour we expect, let's create some global
+        // variables which we use in a variable resolver when we create our engine.
+        // Note: We keep the constants for now, since they are documented.
+        let mut constant_guard = RHAI_GLOBAL_VARIABLES.lock().unwrap();
+        constant_guard.insert("APOLLO_SDL".into(), sdl.to_string().into());
+        constant_guard.insert("APOLLO_START".into(), Instant::now().into());
+        constant_guard.insert(
+            "APOLLO_AUTHENTICATION_JWT_CLAIMS".into(),
+            "apollo_authentication::JWT::claims".to_string().into(),
         );
 
         // Run the AST with our scope to put any global variables
@@ -1258,9 +1262,9 @@ impl Rhai {
         }
 
         // The macro call creates a Rhai module from the plugin module.
-        let module = exported_module!(router_plugin_mod);
+        let module = exported_module!(router_plugin);
 
-        let base64_module = exported_module!(router_base64_mod);
+        let base64_module = exported_module!(router_base64);
 
         // Configure our engine for execution
         engine
@@ -1652,6 +1656,17 @@ impl Rhai {
         // Add common getter/setters for different types
         register_rhai_interface!(engine, supergraph, execution, subgraph);
 
+        // Register a variable resolver.
+        // Note: This API is NOT deprecated, but it is considered volatile and may change in the future.
+        #[allow(deprecated)]
+        engine.on_var(|name, _index, _context| {
+            match name {
+                // Intercept attempts to find "Router" variables and return our "global variables"
+                "Router" => Ok(Some(RHAI_GLOBAL_VARIABLES.lock().unwrap().clone().into())),
+                // Return Ok(None) to continue with the normal variable resolution process.
+                _ => Ok(None),
+            }
+        });
         engine
     }
 
@@ -1667,563 +1682,4 @@ impl Rhai {
 register_plugin!("apollo", "rhai", Rhai);
 
 #[cfg(test)]
-mod tests {
-    use std::str::FromStr;
-
-    use rhai::EvalAltResult;
-    use serde_json::Value;
-    use tower::util::BoxService;
-    use tower::Service;
-    use tower::ServiceExt;
-
-    use super::*;
-    use crate::http_ext;
-    use crate::plugin::test::MockExecutionService;
-    use crate::plugin::test::MockSupergraphService;
-    use crate::plugin::DynPlugin;
-    use crate::services::SubgraphRequest;
-
-    #[tokio::test]
-    async fn rhai_plugin_router_service() -> Result<(), BoxError> {
-        let mut mock_service = MockSupergraphService::new();
-        mock_service
-            .expect_call()
-            .times(1)
-            .returning(move |req: SupergraphRequest| {
-                Ok(SupergraphResponse::fake_builder()
-                    .header("x-custom-header", "CUSTOM_VALUE")
-                    .context(req.context)
-                    .build()
-                    .unwrap())
-            });
-
-        let dyn_plugin: Box<dyn DynPlugin> = crate::plugin::plugins()
-            .find(|factory| factory.name == "apollo.rhai")
-            .expect("Plugin not found")
-            .create_instance_without_schema(
-                &Value::from_str(r#"{"scripts":"tests/fixtures", "main":"test.rhai"}"#).unwrap(),
-            )
-            .await
-            .unwrap();
-        let mut router_service = dyn_plugin.supergraph_service(BoxService::new(mock_service));
-        let context = Context::new();
-        context.insert("test", 5i64).unwrap();
-        let supergraph_req = SupergraphRequest::fake_builder().context(context).build()?;
-
-        let mut supergraph_resp = router_service.ready().await?.call(supergraph_req).await?;
-        assert_eq!(supergraph_resp.response.status(), 200);
-        let headers = supergraph_resp.response.headers().clone();
-        let context = supergraph_resp.context.clone();
-        // Check if it fails
-        let resp = supergraph_resp.next_response().await.unwrap();
-        if !resp.errors.is_empty() {
-            panic!(
-                "Contains errors : {}",
-                resp.errors
-                    .into_iter()
-                    .map(|err| err.to_string())
-                    .collect::<Vec<String>>()
-                    .join("\n")
-            );
-        }
-
-        assert_eq!(headers.get("coucou").unwrap(), &"hello");
-        assert_eq!(headers.get("coming_from_entries").unwrap(), &"value_15");
-        assert_eq!(context.get::<_, i64>("test").unwrap().unwrap(), 42i64);
-        assert_eq!(
-            context.get::<_, String>("addition").unwrap().unwrap(),
-            "Here is a new element in the context".to_string()
-        );
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn rhai_plugin_execution_service_error() -> Result<(), BoxError> {
-        let mut mock_service = MockExecutionService::new();
-        mock_service.expect_clone().return_once(move || {
-            let mut mock_service = MockExecutionService::new();
-            // The execution_service in test.rhai throws an exception, so we never
-            // get a call into the mock service...
-            mock_service.expect_call().never();
-            mock_service
-        });
-
-        let dyn_plugin: Box<dyn DynPlugin> = crate::plugin::plugins()
-            .find(|factory| factory.name == "apollo.rhai")
-            .expect("Plugin not found")
-            .create_instance_without_schema(
-                &Value::from_str(r#"{"scripts":"tests/fixtures", "main":"test.rhai"}"#).unwrap(),
-            )
-            .await
-            .unwrap();
-        let mut router_service = dyn_plugin.execution_service(BoxService::new(mock_service));
-        let fake_req = http_ext::Request::fake_builder()
-            .header("x-custom-header", "CUSTOM_VALUE")
-            .body(Request::builder().query(String::new()).build())
-            .build()?;
-        let context = Context::new();
-        context.insert("test", 5i64).unwrap();
-        let exec_req = ExecutionRequest::fake_builder()
-            .context(context)
-            .supergraph_request(fake_req)
-            .build();
-
-        let mut exec_resp = router_service
-            .ready()
-            .await
-            .unwrap()
-            .call(exec_req)
-            .await
-            .unwrap();
-        assert_eq!(
-            exec_resp.response.status(),
-            http::StatusCode::INTERNAL_SERVER_ERROR
-        );
-        // Check if it fails
-        let body = exec_resp.next_response().await.unwrap();
-        if body.errors.is_empty() {
-            panic!(
-                "Must contain errors : {}",
-                body.errors
-                    .into_iter()
-                    .map(|err| err.to_string())
-                    .collect::<Vec<String>>()
-                    .join("\n")
-            );
-        }
-
-        assert_eq!(
-            body.errors.get(0).unwrap().message.as_str(),
-            "rhai execution error: 'Runtime error: An error occured (line 30, position 5)\nin call to function 'execution_request''"
-        );
-        Ok(())
-    }
-
-    // Some of these tests rely extensively on internal implementation details of the tracing_test crate.
-    // These are unstable, so these test may break if the tracing_test crate is updated.
-    //
-    // This is done to avoid using the public interface of tracing_test which installs a global
-    // subscriber which breaks other tests in our stack which also insert a global subscriber.
-    // (there can be only one...) which means we cannot test it with #[tokio::test(flavor = "multi_thread")]
-    #[test]
-    fn it_logs_messages() {
-        let env_filter = "apollo_router=trace";
-        let mock_writer =
-            tracing_test::internal::MockWriter::new(&tracing_test::internal::GLOBAL_BUF);
-        let subscriber = tracing_test::internal::get_subscriber(mock_writer, env_filter);
-
-        let _guard = tracing::dispatcher::set_default(&subscriber);
-        let engine = Rhai::new_rhai_engine(None);
-        let input_logs = vec![
-            r#"log_trace("trace log")"#,
-            r#"log_debug("debug log")"#,
-            r#"log_info("info log")"#,
-            r#"log_warn("warn log")"#,
-            r#"log_error("error log")"#,
-        ];
-        for log in input_logs {
-            engine.eval::<()>(log).expect("it logged a message");
-        }
-        assert!(tracing_test::internal::logs_with_scope_contain(
-            "apollo_router",
-            "trace log"
-        ));
-        assert!(tracing_test::internal::logs_with_scope_contain(
-            "apollo_router",
-            "debug log"
-        ));
-        assert!(tracing_test::internal::logs_with_scope_contain(
-            "apollo_router",
-            "info log"
-        ));
-        assert!(tracing_test::internal::logs_with_scope_contain(
-            "apollo_router",
-            "warn log"
-        ));
-        assert!(tracing_test::internal::logs_with_scope_contain(
-            "apollo_router",
-            "error log"
-        ));
-    }
-
-    #[test]
-    fn it_prints_messages_to_log() {
-        let env_filter = "apollo_router=trace";
-        let mock_writer =
-            tracing_test::internal::MockWriter::new(&tracing_test::internal::GLOBAL_BUF);
-        let subscriber = tracing_test::internal::get_subscriber(mock_writer, env_filter);
-
-        let _guard = tracing::dispatcher::set_default(&subscriber);
-        let engine = Rhai::new_rhai_engine(None);
-        engine
-            .eval::<()>(r#"print("info log")"#)
-            .expect("it logged a message");
-        assert!(tracing_test::internal::logs_with_scope_contain(
-            "apollo_router",
-            "info log"
-        ));
-    }
-
-    #[tokio::test]
-    async fn it_can_access_sdl_constant() {
-        let dyn_plugin: Box<dyn DynPlugin> = crate::plugin::plugins()
-            .find(|factory| factory.name == "apollo.rhai")
-            .expect("Plugin not found")
-            .create_instance_without_schema(
-                &Value::from_str(r#"{"scripts":"tests/fixtures", "main":"test.rhai"}"#).unwrap(),
-            )
-            .await
-            .unwrap();
-
-        // Downcast our generic plugin. We know it must be Rhai
-        let it: &dyn std::any::Any = dyn_plugin.as_any();
-        let rhai_instance: &Rhai = it.downcast_ref::<Rhai>().expect("downcast");
-
-        let block = rhai_instance.block.load();
-
-        // Get a scope to use for our test
-        let scope = block.scope.clone();
-
-        let mut guard = scope.lock().unwrap();
-
-        // Call our function to make sure we can access the sdl
-        let sdl: String = block
-            .engine
-            .call_fn(&mut guard, &block.ast, "get_sdl", ())
-            .expect("can get sdl");
-        assert_eq!(sdl.as_str(), "");
-    }
-
-    #[test]
-    fn it_provides_helpful_headermap_errors() {
-        let mut engine = Rhai::new_rhai_engine(None);
-        engine.register_fn("new_hm", HeaderMap::new);
-
-        let result = engine.eval::<HeaderMap>(
-            r#"
-    let map = new_hm();
-    map["ümlaut"] = "will fail";
-    map
-"#,
-        );
-        assert!(result.is_err());
-        assert!(matches!(
-            *result.unwrap_err(),
-            EvalAltResult::ErrorRuntime(..)
-        ));
-    }
-
-    // There is a lot of repetition in these tests, so I've tried to reduce that with these two
-    // macros. The repetition could probably be reduced further, but ...
-
-    macro_rules! gen_request_test {
-        ($base: ident, $fn_name: literal) => {
-            let dyn_plugin: Box<dyn DynPlugin> = crate::plugin::plugins()
-                .find(|factory| factory.name == "apollo.rhai")
-                .expect("Plugin not found")
-                .create_instance_without_schema(
-                    &Value::from_str(
-                        r#"{"scripts":"tests/fixtures", "main":"request_response_test.rhai"}"#,
-                    )
-                    .unwrap(),
-                )
-                .await
-                .unwrap();
-
-            // Downcast our generic plugin. We know it must be Rhai
-            let it: &dyn std::any::Any = dyn_plugin.as_any();
-            let rhai_instance: &Rhai = it.downcast_ref::<Rhai>().expect("downcast");
-
-            let block = rhai_instance.block.load();
-
-            // Get a scope to use for our test
-            let scope = block.scope.clone();
-
-            let mut guard = scope.lock().unwrap();
-
-            // We must wrap our canned request in Arc<Mutex<Option<>>> to keep the rhai runtime
-            // happy
-            let request = Arc::new(Mutex::new(Some($base::fake_builder().build())));
-
-            // Call our rhai test function. If it return an error, the test failed.
-            let result: Result<(), Box<rhai::EvalAltResult>> =
-                block
-                    .engine
-                    .call_fn(&mut guard, &block.ast, $fn_name, (request,));
-            result.expect("test failed");
-        };
-    }
-
-    macro_rules! gen_response_test {
-        ($base: ident, $fn_name: literal) => {
-            let dyn_plugin: Box<dyn DynPlugin> = crate::plugin::plugins()
-                .find(|factory| factory.name == "apollo.rhai")
-                .expect("Plugin not found")
-                .create_instance_without_schema(
-                    &Value::from_str(
-                        r#"{"scripts":"tests/fixtures", "main":"request_response_test.rhai"}"#,
-                    )
-                    .unwrap(),
-                )
-                .await
-                .unwrap();
-
-            // Downcast our generic plugin. We know it must be Rhai
-            let it: &dyn std::any::Any = dyn_plugin.as_any();
-            let rhai_instance: &Rhai = it.downcast_ref::<Rhai>().expect("downcast");
-
-            let block = rhai_instance.block.load();
-
-            // Get a scope to use for our test
-            let scope = block.scope.clone();
-
-            let mut guard = scope.lock().unwrap();
-
-            // We must wrap our canned response in Arc<Mutex<Option<>>> to keep the rhai runtime
-            // happy
-            let response = Arc::new(Mutex::new(Some($base::default())));
-
-            // Call our rhai test function. If it return an error, the test failed.
-            let result: Result<(), Box<rhai::EvalAltResult>> =
-                block
-                    .engine
-                    .call_fn(&mut guard, &block.ast, $fn_name, (response,));
-            result.expect("test failed");
-        };
-    }
-
-    #[tokio::test]
-    async fn it_can_process_supergraph_request() {
-        let dyn_plugin: Box<dyn DynPlugin> = crate::plugin::plugins()
-            .find(|factory| factory.name == "apollo.rhai")
-            .expect("Plugin not found")
-            .create_instance_without_schema(
-                &Value::from_str(
-                    r#"{"scripts":"tests/fixtures", "main":"request_response_test.rhai"}"#,
-                )
-                .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        // Downcast our generic plugin. We know it must be Rhai
-        let it: &dyn std::any::Any = dyn_plugin.as_any();
-        let rhai_instance: &Rhai = it.downcast_ref::<Rhai>().expect("downcast");
-
-        let block = rhai_instance.block.load();
-
-        // Get a scope to use for our test
-        let scope = block.scope.clone();
-
-        let mut guard = scope.lock().unwrap();
-
-        // We must wrap our canned request in Arc<Mutex<Option<>>> to keep the rhai runtime
-        // happy
-        let request = Arc::new(Mutex::new(Some(
-            SupergraphRequest::canned_builder()
-                .operation_name("canned")
-                .build()
-                .expect("build canned supergraph request"),
-        )));
-
-        // Call our rhai test function. If it return an error, the test failed.
-        let result: Result<(), Box<rhai::EvalAltResult>> = block.engine.call_fn(
-            &mut guard,
-            &block.ast,
-            "process_supergraph_request",
-            (request,),
-        );
-        result.expect("test failed");
-    }
-
-    #[tokio::test]
-    async fn it_can_process_execution_request() {
-        gen_request_test!(ExecutionRequest, "process_execution_request");
-    }
-
-    #[tokio::test]
-    async fn it_can_process_subgraph_request() {
-        gen_request_test!(SubgraphRequest, "process_subgraph_request");
-    }
-
-    #[tokio::test]
-    async fn it_can_process_supergraph_response() {
-        gen_response_test!(RhaiSupergraphResponse, "process_supergraph_response");
-    }
-
-    #[tokio::test]
-    async fn it_can_process_supergraph_deferred_response() {
-        gen_response_test!(
-            RhaiSupergraphDeferredResponse,
-            "process_supergraph_response"
-        );
-    }
-
-    #[tokio::test]
-    async fn it_can_process_execution_response() {
-        gen_response_test!(RhaiExecutionResponse, "process_execution_response");
-    }
-
-    #[tokio::test]
-    async fn it_can_process_execution_deferred_response() {
-        gen_response_test!(RhaiExecutionDeferredResponse, "process_execution_response");
-    }
-
-    #[tokio::test]
-    async fn it_can_process_subgraph_response() {
-        let dyn_plugin: Box<dyn DynPlugin> = crate::plugin::plugins()
-            .find(|factory| factory.name == "apollo.rhai")
-            .expect("Plugin not found")
-            .create_instance_without_schema(
-                &Value::from_str(
-                    r#"{"scripts":"tests/fixtures", "main":"request_response_test.rhai"}"#,
-                )
-                .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        // Downcast our generic plugin. We know it must be Rhai
-        let it: &dyn std::any::Any = dyn_plugin.as_any();
-        let rhai_instance: &Rhai = it.downcast_ref::<Rhai>().expect("downcast");
-
-        let block = rhai_instance.block.load();
-        // Get a scope to use for our test
-        let scope = block.scope.clone();
-
-        let mut guard = scope.lock().unwrap();
-
-        // We must wrap our canned response in Arc<Mutex<Option<>>> to keep the rhai runtime
-        // happy
-        let response = Arc::new(Mutex::new(Some(subgraph::Response::fake_builder().build())));
-
-        // Call our rhai test function. If it return an error, the test failed.
-        let result: Result<(), Box<rhai::EvalAltResult>> = block.engine.call_fn(
-            &mut guard,
-            &block.ast,
-            "process_subgraph_response",
-            (response,),
-        );
-        result.expect("test failed");
-    }
-
-    #[test]
-    fn it_can_urlencode_string() {
-        let engine = Rhai::new_rhai_engine(None);
-        let encoded: String = engine
-            .eval(r#"urlencode("This has an ümlaut in it.")"#)
-            .expect("can encode string");
-        assert_eq!(encoded, "This%20has%20an%20%C3%BCmlaut%20in%20it.");
-    }
-
-    #[test]
-    fn it_can_urldecode_string() {
-        let engine = Rhai::new_rhai_engine(None);
-        let decoded: String = engine
-            .eval(r#"urldecode("This%20has%20an%20%C3%BCmlaut%20in%20it.")"#)
-            .expect("can decode string");
-        assert_eq!(decoded, "This has an ümlaut in it.");
-    }
-
-    #[test]
-    fn it_can_base64encode_string() {
-        let engine = Rhai::new_rhai_engine(None);
-        let encoded: String = engine
-            .eval(r#"base64::encode("This has an ümlaut in it.")"#)
-            .expect("can encode string");
-        assert_eq!(encoded, "VGhpcyBoYXMgYW4gw7xtbGF1dCBpbiBpdC4=");
-    }
-
-    #[test]
-    fn it_can_base64decode_string() {
-        let engine = Rhai::new_rhai_engine(None);
-        let decoded: String = engine
-            .eval(r#"base64::decode("VGhpcyBoYXMgYW4gw7xtbGF1dCBpbiBpdC4=")"#)
-            .expect("can decode string");
-        assert_eq!(decoded, "This has an ümlaut in it.");
-    }
-
-    async fn base_process_function(fn_name: &str) -> Result<(), Box<rhai::EvalAltResult>> {
-        let dyn_plugin: Box<dyn DynPlugin> = crate::plugin::plugins()
-            .find(|factory| factory.name == "apollo.rhai")
-            .expect("Plugin not found")
-            .create_instance_without_schema(
-                &Value::from_str(
-                    r#"{"scripts":"tests/fixtures", "main":"request_response_test.rhai"}"#,
-                )
-                .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        // Downcast our generic plugin. We know it must be Rhai
-        let it: &dyn std::any::Any = dyn_plugin.as_any();
-        let rhai_instance: &Rhai = it.downcast_ref::<Rhai>().expect("downcast");
-
-        let block = rhai_instance.block.load();
-
-        // Get a scope to use for our test
-        let scope = block.scope.clone();
-
-        let mut guard = scope.lock().unwrap();
-
-        // We must wrap our canned response in Arc<Mutex<Option<>>> to keep the rhai runtime
-        // happy
-        let response = Arc::new(Mutex::new(Some(subgraph::Response::fake_builder().build())));
-
-        // Call our rhai test function. If it doesn't return an error, the test failed.
-        block
-            .engine
-            .call_fn(&mut guard, &block.ast, fn_name, (response,))
-    }
-
-    #[tokio::test]
-    async fn it_can_process_om_subgraph_forbidden() {
-        if let Err(error) = base_process_function("process_subgraph_response_om_forbidden").await {
-            let processed_error = process_error(error);
-            assert_eq!(processed_error.status, StatusCode::FORBIDDEN);
-            assert_eq!(processed_error.message, "I have raised a 403");
-        } else {
-            // Test failed
-            panic!("error processed incorrectly");
-        }
-    }
-
-    #[tokio::test]
-    async fn it_can_process_string_subgraph_forbidden() {
-        if let Err(error) = base_process_function("process_subgraph_response_string").await {
-            let processed_error = process_error(error);
-            assert_eq!(processed_error.status, StatusCode::INTERNAL_SERVER_ERROR);
-            assert_eq!(processed_error.message, "rhai execution error: 'Runtime error: I have raised an error (line 124, position 5)\nin call to function 'process_subgraph_response_string''");
-        } else {
-            // Test failed
-            panic!("error processed incorrectly");
-        }
-    }
-
-    #[tokio::test]
-    async fn it_cannot_process_ok_subgraph_forbidden() {
-        if let Err(error) = base_process_function("process_subgraph_response_om_ok").await {
-            let processed_error = process_error(error);
-            assert_eq!(processed_error.status, StatusCode::INTERNAL_SERVER_ERROR);
-            assert_eq!(processed_error.message, "I have raised a 200");
-        } else {
-            // Test failed
-            panic!("error processed incorrectly");
-        }
-    }
-
-    #[tokio::test]
-    async fn it_cannot_process_om_subgraph_missing_message() {
-        if let Err(error) =
-            base_process_function("process_subgraph_response_om_missing_message").await
-        {
-            let processed_error = process_error(error);
-            assert_eq!(processed_error.status, StatusCode::BAD_REQUEST);
-            assert_eq!(processed_error.message, "rhai execution error: 'Runtime error: #{\"status\": 400} (line 135, position 5)\nin call to function 'process_subgraph_response_om_missing_message''");
-        } else {
-            // Test failed
-            panic!("error processed incorrectly");
-        }
-    }
-}
+mod tests;
