+++ apollo-router/src/plugins/rhai/subgraph.rs
@@ -1,0 +1,3 @@
+//! subgraph module
+
+pub(crate) use crate::services::subgraph::*;
