+++ apollo-router/src/plugins/rhai/supergraph.rs
@@ -1,0 +1,5 @@
+//! supergraph module
+
+pub(crate) use crate::services::supergraph::*;
+pub(crate) type Response = super::RhaiSupergraphResponse;
+pub(crate) type DeferredResponse = super::RhaiSupergraphDeferredResponse;
